+++ src/game/effect_support.rs
@@ -820,6 +820,7 @@ impl Game {
             | ObjectPredicateDef::ToughnessExactly(_)
             | ObjectPredicateDef::ToughnessLessThan(_)
             | ObjectPredicateDef::PowerGreaterThan(_)
+            | ObjectPredicateDef::PowerLessThan(_)
             | ObjectPredicateDef::ToughnessGreaterThan(_)
             | ObjectPredicateDef::ControlledBy(_)
             | ObjectPredicateDef::Supertype(_)